requirements.txt
appdirs==1.4.4
apturl==0.5.2
attrs==23.2.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.8.2
bottle==0.12.25
Brlapi==0.8.3
build==1.2.1
buildozer==1.5.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
Cython==3.0.10
dbus-python==1.2.18
defer==1.0.6
deluge==2.0.3
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
duplicity==0.8.21
exceptiongroup==1.2.1
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==3.0.3
future==0.18.2
GeoIP==1.3.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isodate==0.6.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libtorrent===2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
protobuf==3.12.4
proxy-tools==0.1.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycountry==24.6.1
pycryptodome==3.20.0
pycups==2.0.1
pygame==2.1.2
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-vlc==3.0.20123
pytz==2022.1
pywebview==5.1
pyxdg==0.27
PyYAML==5.4.1
rencode==1.0.6
reportlab==3.6.8
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.2.2
sh==1.14.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.1
streamlink==6.7.4
systemd-python==234
tk==0.1.0
toml==0.10.2
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==22.1.0
typing_extensions==4.12.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.2
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0



